requirements.txt
Django==4.2.7
djangorestframework==3.15.1 
djangorestframework-simplejwt==5.3.1
django-rest-passwordreset==1.3.0
Pillow==10.3.0 
python-decouple==3.8
django-cors-headers==4.3.0
drf-yasg==1.21.7 
httpx==0.25.2
PyJWT>=2.8.0,<3.0.0
tzdata
django-allauth>=0.58.2 
dj-database-url>=2.1.0 
psycopg2-binary>=2.9.9